requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
typing-extensions==4.8.0